minor cleanup, mod still needs to be restructured + replaceInSource script

+++ format.sh
@@ -3,7 +3,9 @@ shopt -s globstar
 
 echo "if this fails please look in the script for requirements"
 ##sudo apt install cargo ##debian
-cargo install --root . stylua --features lua52
+if ! test -f "./bin/stylua"; then
+	cargo install --root . stylua --features lua52
+fi
 
 ./bin/stylua --respect-ignores --verbose ./**/*.lua
 ##--respect-ignores to respect ignores in .styluaignore, factorios lua is a bit modified shit happens
